db fix

+++ migrations/2025-05-03-222225_create_product/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE Products (
   id serial primary key,
   title varchar(100) not null,
-  price float not null,
+  price money not null,
   description text not null,
   image text
 );
